requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
circuitbreaker==2.1.3
click==8.0.4
cryptography==44.0.2
datasets==3.5.0
dill==0.3.8
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
jmespath==0.10.0
multidict==6.2.0
multiprocess==0.70.16
numpy==2.2.4
packaging==24.2
pandas==2.2.3
prompt-toolkit==3.0.43
propcache==0.3.1
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
terminaltables==3.1.10
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.1
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
pillow==11.2.1
scikit-learn==1.6.1
git+https://github.com/benjaminkost/pythreshold # forked pythreshold==0.3.1
git+https://github.com/benjaminkost/Contour-Detection # forked Contour-Detection